requirements.txt
asgiref==3.4.1
astroid==2.7.3
Django==3.2.7
django-cors-headers==3.8.0
django-filter==2.4.0
djangorestframework==3.12.4
importlib-metadata==4.8.1
isort==5.9.3
lazy-object-proxy==1.6.0
Markdown==3.3.4
mccabe==0.6.1
pkg-resources==0.0.0
platformdirs==2.3.0
pylint==2.10.2
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.2
wrapt==1.12.1
zipp==3.5.0
